#519338 moved redirects service into sitecore-jss
