+++ packages/sitecore-jss/src/site/index.ts
@@ -3,3 +3,9 @@ export {
   GraphQLRobotsService,
   GraphQLRobotsServiceConfig,
 } from './graphql-robots-service';
+export {
+  RedirectType,
+  RedirectsQueryResult,
+  GraphQLRedirectService,
+  GraphQLRedirectsServiceConfig,
+} from './graphql-redirect-service';
